+++ graphix/qasm3_exporter.py
@@ -2,14 +2,13 @@
 
 from __future__ import annotations
 
-from math import pi
 from typing import TYPE_CHECKING
 
 # assert_never added in Python 3.11
 from typing_extensions import assert_never
 
 from graphix.command import CommandKind
-from graphix.fundamentals import Axis, Plane
+from graphix.fundamentals import Axis, ParameterizedAngle, Plane, angle_to_rad
 from graphix.instruction import Instruction, InstructionKind
 from graphix.pretty_print import OutputFormat, angle_to_str
 from graphix.states import BasicStates, State
@@ -19,7 +18,6 @@ if TYPE_CHECKING:
 
     from graphix import Circuit, Pattern
     from graphix.command import Command
-    from graphix.parameter import ExpressionOrFloat
 
 
 def circuit_to_qasm3(circuit: Circuit) -> str:
@@ -64,12 +62,11 @@ def qasm3_gate_call(gate: str, operands: Iterable[str], args: Iterable[str] | No
     return f"{gate}({args_str}) {operands_str}"
 
 
-def angle_to_qasm3(angle: ExpressionOrFloat) -> str:
+def angle_to_qasm3(angle: ParameterizedAngle) -> str:
     """Get the OpenQASM3 representation of an angle."""
     if not isinstance(angle, float):
         raise TypeError("QASM export of symbolic pattern is not supported")
-    rad_over_pi = angle / pi
-    return angle_to_str(rad_over_pi, output=OutputFormat.ASCII, multiplication_sign=True)
+    return angle_to_str(angle, output=OutputFormat.ASCII, multiplication_sign=True)
 
 
 def instruction_to_qasm3(instruction: Instruction) -> str:
@@ -204,12 +201,13 @@ def command_to_qasm3_lines(cmd: Command) -> Iterator[str]:
         if cmd.plane == Plane.XY:
             yield f"h q{cmd.node};\n"
         if cmd.angle != 0:
+            rad_angle = angle_to_rad(cmd.angle)
             if cmd.plane == Plane.XY:
-                yield f"rx({-cmd.angle * pi}) q{cmd.node};\n"
+                yield f"rx({-rad_angle}) q{cmd.node};\n"
             elif cmd.plane == Plane.XZ:
-                yield f"ry({-cmd.angle * pi}) q{cmd.node};\n"
+                yield f"ry({-rad_angle}) q{cmd.node};\n"
             else:
-                yield f"rx({cmd.angle * pi}) q{cmd.node};\n"
+                yield f"rx({rad_angle}) q{cmd.node};\n"
         yield f"bit c{cmd.node};\n"
         yield f"c{cmd.node} = measure q{cmd.node};\n"