+++ tests/test_qasm3_exporter_to_qiskit.py
@@ -116,6 +116,7 @@ def test_to_qasm3_random_circuit(fx_bg: PCG64, jumps: int) -> None:
     depth = 5
     circuit = rand_circuit(nqubits, depth, rng=rng)
     pattern = circuit.transpile().pattern
+    pattern.remove_input_nodes()
     pattern.perform_pauli_measurements()
     pattern.minimize_space()